refactor: restructure header code
